MegaClient::putfa(...) now returns an error code rather than a boolean.
This commit solves a problem that I introduced earlier that caused a
crash. The crash occured because my changes always invoke the
application's putfa_result(...) callback in an error code such as when
an FA is too large.

The crash occurs because the IL's request procesing code expects that
any immediate errors do not deallocate the request but invoking the IL's
putfa_result(...) callback *does* deallocate the request.

So, the solution here is simply to not invoke the application's callback
on immediate failure (EARGS, EKEY) and instead, return an error code
that we can use to signal the correct failure reason.

The commit also fixes some indentation issues.