Retrieve and notify storage status upon changes
Upon changes of `^!usl` user's attribute (storage state), the SDK used
to notify the `STORAGE_STATE_CHANGE`. This commit will skip such
notification and the need of the app to explicitly request the account
details (`uq` command), in order to get the current storage state.

Instead, the SDK will fetch the value and notify accordingly upon
response of `uga` command.