Remove legacy functionality in MegaApi::getPasswordManagerBase
Before MegaApi introduced ClientType parameter,
MegaApi::getPasswordManagerBase was reponsible of creating the
Password Manager Base in case it didn't exist (first time a user would
start Password Manager app).

This functionality is now part of fetch nodes operation which is
required to be successfully performed to be able to access any node in
the account.