SDK-4535. Reused or obsolete FSIDs result in mixing different files
FSIDs are used to detect moves by matching a current FSID with a cached
FSID.

However, they can be reused due to filesystem operations: delete and
create files in short time or even updating files by replacement can end
up with new files reusing recently freed FSIDs.

For that reason, comparing fingerprints is necessary to ensure that
matched nodes by FSID are really the same moved file and not two
different files with reused FSIDs, leading to mixing unrelated files
without the user being aware.

This commit refactors Syncs::findLocalNodeBySyncedFsid and
Syncs::findLocalNodeByScannedFsid, defining two new methods:

1. findLocalNodeByFsid() -> method with the general
logic to match by fsid.
2. areNodesMatchedByFsidEquivalent() -> method with the core
logic to consider two nodes matched by FSID as equivalent, including
fingerprint match.

The initial reason for removing the fingerprint comparison for two nodes
matching by FSID was to avoid a corner case: a file which has been
modified and moved right after putnodes started, but wasn't completed.

In that scenario, the moved file would match with the cached localnode
FSID, but would have a fingerprint mismatch.

If the putnodes is detected for the source localnode and the fingerprint
of the upload matches with the target node, the sync will wait for
putnodes completion, keeping the current behavior for that scenario.