Expanded logic for Sync Uploads Throttling
1. Added public interfaces to retrieve configurable throttling values,
set configurable values and get the defined lower and upper limits
2. Create an interface IUploadThrottlingManager which is the base class
of UploadThrottlingManager
3. Add public methods in Syncs to access the mUploadThrottlingManager
functionality, use those methods to check that the thread is the
expected one and other Syncs-related checks.
4. Add also the necessary async methods in Syncs to change or get those
values by enqueueing the tasks to be executed on the sync thread.
5. Readapt documentation as needed.
6. Add pending const modifiers.
7. Remove Syncs reference from UploadsThrottlingManager and move the
final queueClient logic to a specific method from Syncs