Make sure our tag commands properly report failures.
This commit makes sure that our tag commands properly return an error
code from the client when an error can be detected immediately.

Prior to this change, we would only report an error if a request was
sent to the backend and it responded with an error. This works most of
the time but fails because the client can detect error cases immediately
in some situations. If that happened, the user's callback would never be
invoked and so any waiters would wait forever.