Changed nodeTreeMutex to a recursive mutex.
This has been a change I've been wanting to make for some time, simply
to avoid having to introduce lock-free variants of variouo functions.

The reason this is useful is that there are times where FUSE code needs
to call the client to retrieve information about a node and the caller
itself is unable to determine whether the nodeTreeMutex is held or not.