Added the SelectFirst functor.
This'll be useful later on when we implement the tree.

The reason behind this is that we want to be able to use the tree to
implement both a map and set.

When the tree is being used to implement a set, each node's value will
be used directly as the key.

When the tree is being used to implement a map, each node's value will
be a pair and we need some way to retrieve the key value from that pair.