Remove MegaFileSystemAccess entity
`MegaFileSystemAccess` is designed as a compile-time class definition
to be platform specific. However, Android platform may require
different entities depending on wether Android's `FileWrapper` JNI is
present and available or not.

In order to allow runtime changes for this required `FileSystemAccess`
the design changes of this commit are required.

No functional changes are introduced in this commit.