fix: sync tests failures
1. waitonSyncs now take name clashes into account along with
stall issues for early exit. Name clashes were not considered
before and that caused long execution times for sync tests
with name clashes (5 time per call to waitOnSyncs).
2. waitOnSyncs and waitFor now take into account the time
lapsed for client doWait() calls, increasing the max timeout
with the this value. The impact is negiglible for most cases,
but for the rare cases when doWait() duration takes several
seconds just when the waitOnSyncs or waitFor is called, it
caused a failure in the conditions waiting to match when the
timeout was small (ex: 8 secs timeout for waitFor(), dowait
takes 7 secs, so there is only 1 sec left for the sync
operations needed, which may need just 1.5 secs, which is a
short time but above the timeout).