Fix transfer localfilename un/serialization for macos
The way we were Un/serializing transfer localfilename for Macos wasn't
symmetric, so automated test SyncPaths was failing.

The reason behind this failure was that recently we have introduced new
code to serialize localfilename as a LocalPath. This new code called
path2local which contains specific implementation for macos
(path2localMac) but not on the other way.

The solution to this problem has been avoid performing any conversion
format in strings (UTF8 to Windows Unicode or viceversa), upon
Un/serializing so we can stop calling LocalPath::local2path and
LocalPath::path2local.