Review: Rename mKeyPointer to mValuePointer.
This commit also renames a bunch of stuff related to mKeyPointer.

The rationale for this change is that mKeyPointer isn't necessarily
pointing to a key in a given node.

An example would be when we're using an AVL tree to implement a map. In
this case, each node will contain a std::pair<X, Y> value where X is the
key and Y is the value that the key maps to.

mValuePointer will tell the tree how to access this pair.
Traits::KeyFunction tells the tree how to extract the key from that pair.

KeyIsEqualityComparableV has been renamed to ValueIsEqualityComparableV.

The new name is correct because when we compare two trees, we aren't
just checking that all the keys are the same, we need to check that all
the values in every node are the same.