test: relaxed predicate for waitForSyncTotalStallsStateUpdate(true)
If waitFor(SyncTotalStallsStateUpdate(true)) is false, it
could be due to a change of stall state that was temporary
false due to pending scans, and then set to true again when
there only was 1 stall.

For that case, checking that the stall state is true should be
fine as a relaxed predicate condition, complemented by
checking the number of detected stalls afterwards.