test: fix race condition in MirroringInternalBackupResumesInMirroringMode
Once setupSync_mainthread is completed and the sync is
running, but before the whole function returns, the initial
file may start the upload and the mOnFileAdded callback would
be executed.

At that point, the backupId is not yet returned and checking
syncConfigByBackupId would fail.

This commit fixes that race condition.