fix: avoid race conditions when processing syncThreadActions
Syncs::syncThreadActions is of type ThreadSafeDeque. All its
methods to push and pop functions are thread safe. However,
processing it in the sync loop with a while() loop was not
100% thread safe: a function could be added when the "while"
was running, and that new function could depend on some output
that should have been processed later in the sync loop as a
consequence of processing the first function.

Hence, a second queued function added before completing the
sync loop iteration when the first queued was processed, would
be processed in the same while() loop iteration, and the
expectations could fail because it depended on the first
function that was processed in the while() loop but the
consequences (ex: cleaning/deleting the sync that has been
disabled) were not, causing unexpected scenarios.

This commit fixes that situation by creating a copy to be
processed in the sync loop, avoiding this kind of race
conditions.