Tweak the f(raw)?read(...), fwrite(...) and sysread(...) functions.
All of these functions now accept an optional "retry" parameter which is
set by the function to indicate whether an encountered error is
transient. The reason why this is useful is that it allows us to store
that retry state outside of a FileAccess instance which makes life a lot
nicer when we're sharing such an instance between threads.

fwrite(...) also accepts an optional "numWritten" parameter which tracks
how many bytes of data were actually written in the case of a failure.