fix: restore the original name of a SyncUpload transfer that has been stolen-and-restored during a move
1. When a fileA is renamed (ex: to fileB) the sync engine
handles a move operation that can last several seconds.
2. When, during that move operation, a new fileA is created,
with different contents or fingerprint but same name, a new
SyncUpload transfer is created to sync the new contents of
fileA.
3. A previous bug caused the SyncUpload transfer for new fileA
to be spuriously moved to fileB, due to a mistake of the move
operation (considered that the new fileA is still the
old/renamed fileA).
4. The fix for that bug didn't consider that the "steal"
operation of the SyncUpload also included changing the path of
the SyncUpload. The SyncUpload was returned to the rightful
owner (new fileA) but the name of the file, used for the
transfer, remained with the name of the illegitimate owner
(fileB).

This commit fixes that scenario.