Store current file system name for Macos at fopen
Fix issue at posix file system (case insensitive) when renaming a file
to a name that differs only by case (e.g., "B" → "A" when "a" already
exists) does not actually change the case of the targe filename.
Instead, the existing "a" is overwritten in content but its original
casing is preserved.
Add paramter to downloadNode to indicate if path reference a folder or
not