Many Client(Adapter)? tweaks.
The goal of this commit was to make using the Client(Adapter)? classes a
little more ergonomic.

The only changes that really affect any users are the fact that both
childNames(...) and handle(...) now return an ErrorOr<T> instead of some
value directly.

This is useful because it lets client code understand why they couldn't
get the handle of the node.

The commit also includes some minor tweaks to get(...) so that we can
easily tell whether the lookup failed because the parent doesn't exist
or because the child itself doesn't exist.