Modify findCloneNodeCandidate to perform Meta Mac comparison
Prior to this change the findCloneNodeCandidate method was not accurate
enough due to comparissons just based on fingerprint. With this change
we will calculate file Meta Mac and will compare with NodeKey with give
us an exact result.

Add new sync upload operations test

Also add test DuplicatedFilesUploadDifferentMtime

This test tries to exercise the use case where we have a file with same
content and size but different mtime. With current code it means that
fingerprint won't match with any cloud candidate as mtime is considered
as part of fingerprint.