Remove support for coexistence of API v2 and v3 commands
The SDK already sends all commands by using the APIv3, we don't need to
maintain the ability to split the queue of commands into independent
requests depending on the version of the command.

This commit also restores part of the
SdkTestShare::TestSharesContactVerification that was disabled due to a
race between command's response and action packets, which doesn't happen
in APIv3.