Add virtual column to nodes table for fingerprint withouth mtime
In case we want to upload/sync a file that already exist in cloud in
same parent node and with same name and META_MAC, but with different
mtime we just can update mtime for that node instead of performing a
full upload or a clone node.

For this purpose we need to search nodes by FP in DB excluding mtime
with a good performance.

Adding a virtual column and a new index it's the best approach to
perform these kind of queries without affecting performance.