FileReadRequestSet is now a multiset instead of a set.
This fixes a nasty bug found by Jeff where multiple pending read
requests to the same range weren't being correctly tracked.

The reason is obvious when you look at it: The set is keyed on the read
request's range and a set can only contain one instance of a particular
key.