test: excercise syncs transfers with same mac and mtime only change
- Adapt SyncFingerprintCollisionTest.DifferentMacSameName
- Change contents of the file with CRC collision, keeping same
mtime: sync won't do anything as fingerprint is the same. MAC
differs but it won't be calculated as mtime is the same.
- Change mtime of the file, fingerprint is now the same for
CRC and size only, MAC calculation will be performed. It
differs, so the file will be re-uploaded.
- Change mtime of the file again, now files have same CRC and
same size, different mtime but same MAC. Client2 should not
download the file but update the local file mtime.